[RENAME - #92] DTO 클래스들을 각 Layer 에 맞게 분리
