+++ src/main/java/com/dateplan/dateplan/domain/anniversary/controller/dto/response/AnniversaryDatesResponse.java
@@ -1,5 +1,6 @@
-package com.dateplan.dateplan.domain.anniversary.dto;
+package com.dateplan.dateplan.domain.anniversary.controller.dto.response;
 
+import com.dateplan.dateplan.domain.anniversary.service.dto.response.AnniversaryDatesServiceResponse;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
@@ -10,7 +11,7 @@ import lombok.Getter;
 @Builder
 public class AnniversaryDatesResponse {
 
-	private List<LocalDate> anniversaryDates = new ArrayList<>();
+	private List<LocalDate> anniversaryDates;
 
 	public static AnniversaryDatesResponse from(AnniversaryDatesServiceResponse serviceResponse) {
